Iteration function for Tree-based MM

Function that leverages xboosting() to estimate the trees. predict.xgb() for prediction, and mem_boost_gll() for convergence. We first introduce the general unit-level formula for the Extreme Boosting Mix Effect Model $$y_i = f(X_i) + Z_i b_i + e_i$$ Where: \(f(X_i)\) represents the Extreme Gradient Boosting Model. \(Z_i\) is an \(n_i \times k\) matrix for the random effects. \(b_i\) is a vector of random effects. \(e_i\) are the error terms.

We assume that the current estimates of the random effects are correct, that is, \(\hat{b}_i = b_i\), and a tree is fit to a transformed outcome from which the random effects have been removed $$\bar{y}_i = y_i - Z_i \hat{b}_i$$

Once \(\hat{f(X_i)}\) has been fit, random effects \(\hat{b}_i\) are updated.

$$\hat{b}_i = \hat{D} Z_i^T \hat{V}_i^{-1} [y_i - \hat{f}(X_i)]$$

For formula reference for the generalised log-likelihood (GLL) criterion please refer to mem_boost_gll

boost_mem(
  formula,
  data = NULL,
  random = NULL,
  shrinkage = 0.3,
  loss = "reg:squarederror",
  interaction.depth = 20,
  n.trees = 100,
  minsplit = 20,
  subsample = 0.5,
  lambda = 1,
  alpha = 0,
  conv_memboost = 0.001,
  maxIter_memboost = 100,
  minIter_memboost = 0,
  verbose_memboost = FALSE
)

Arguments

formula

an object of class formula

data

list or environment (or object coercible by as.data.frame to a data frame) containing the variables in the model.

random

an object of class formula with the random intercept

shrinkage

learning rate. Default: 0.1

loss

by default it uses "reg:squarederror" more functions available in the xgboost documentation. Users, can pass a self-defined function to it.

interaction.depth

maximum depth of the trees. Default: 1

n.trees

number of trees to be generated. Default: 100

minsplit

minimum observations for a tree to split. Default: 20

subsample

sub-sample size for the tree training. Default: 0.5

lambda

regularization term on weights. Default: 1

alpha

regularization term on weights. Default: 0

conv_memboost

Convergence threshold. Default: 0.001

maxIter_memboost

maximum iterations. Default: 100

minIter_memboost

minimum iterations. Default: 0

verbose_memboost

Print information

Value

list of values

References

Marie Salditt, Sarah Humberg & Steffen Nestler (2023) Gradient Tree Boosting for Hierarchical Data, Multivariate Behavioral Research, 58:5, 911-937, DOI: 10.1080/00273171.2022.2146638