Computes Extreme Gradient Boosting Trees

Relies on the Extreme Gradient Boosting model to generate the decision trees and subsequently an assembly of m trees.

Initialize model with a constant value: $$\hat{f}_0(x) = \arg \min \sum_{i=1}^N L(y_i, \theta)$$

Compute gradients and hessians.

fit a weak learner:

$$\hat{\phi}_m = \sum_{i =1}^{Im} \phi_{im}I(x \in R_{im})$$

Update model:

$$\hat{f}_{m-1}(x) + \eta \cdot \hat{\phi}_m$$

xboosting(
  formula,
  data = NULL,
  loss = "reg:squarederror",
  n.trees = 100,
  shrinkage = 0.1,
  interaction.depth = 1,
  minsplit = 20,
  lambda = 1,
  alpha = 0,
  subsample = 0.5
)

Arguments

formula

an object of class formula

data

list or environment (or object coercible by as.data.frame to a data frame) containing the variables in the model.

loss

by default it uses "reg:squarederror" more functions available in the xgboost documentation. Users, can pass a self-defined function to it.

n.trees

number of trees to be generated. Default: 100

shrinkage

learning rate. Default: 0.1

interaction.depth

maximum depth of the trees. Default: 1

minsplit

minimum observations for a tree to split. Default: 20

lambda

regularization term on weights. Default: 1

alpha

regularization term on weights. Default: 0

subsample

sub-sample size for the tree training: Default: 0.5

Value

an xtremeBoost object

References

Chen T, He T, Benesty M, Khotilovich V, Tang Y, Cho H, Chen K, Mitchell R, Cano I,Zhou T, Li M, Xie J, Lin M, Geng Y, Li Y, Yuan J (2024). _xgboost: Extreme Gradient Boosting_. R package version 1.7.7.1,<https://CRAN.R-project.org/package=xgboost>.

Marie Salditt, Sarah Humberg & Steffen Nestler (2023) Gradient Tree Boosting for Hierarchical Data, Multivariate Behavioral Research, 58:5, 911-937, DOI: 10.1080/00273171.2022.2146638

Corral, P., Henderson, H., & Segovia, S. (2023). Poverty Mapping in the Age of Machine Learning. World Bank policy research working paper. https://doi.org/10.1596/1813-9450-10429